+++ wls-gui-wahllokalsystem/src/components/dse/BaseStimmzettelUebersichtTable.vue
